requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
